+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsDiagnosticTestGenerated.java
@@ -1335,6 +1335,12 @@ public class JsDiagnosticTestGenerated extends AbstractJsDiagnosticTest {
     }
 
     @Test
+    @TestMetadata("DuplicateEffectiveMapKeys.kt")
+    public void testDuplicateEffectiveMapKeys() {
+      run("DuplicateEffectiveMapKeys.kt");
+    }
+
+    @Test
     @TestMetadata("DuplicateMapKeys.kt")
     public void testDuplicateMapKeys() {
       run("DuplicateMapKeys.kt");